Bug-ridden base integration
Algorithm
+Goes from start to end!
-Ignores obstacles
-Crashes if start/end too distant
-Sometimes thinks it's blocked when it isn't
